hgc-ml/requirements.txt
mlflow==2.11.3
snowflake-connector-python==3.8.0
snowflake-sqlalchemy==1.5.1
pandas==2.2.1
scikit-learn==1.4.1.post1
xgboost==2.0.3
jupyterlab==4.1.5
python-dotenv==1.0.1
